Fix support for titles with symbols in them
The shell code here was all wrong and unsafe. This fixes the shell
argument passing so that all kinds of symbols and characters are
allowed.

The "*" -> "_" replacement isn't strictly necessary but it's usually
polite not to put *'s in your file names

Test with: jbripper.py u p spotify:track:4XoP1AkbOurU9CeZ2rMEz2